requirements.txt
asgiref==3.4.1
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
psycopg==3.0.4
psycopg2-binary==2.9.2
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
